factor out ScalarParameterControl
